requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Async MongoDB driver
motor==3.6.0
dnspython==2.7.0

# Async HTTP client
httpx==0.27.2

# Settings management
pydantic-settings==2.5.2

# Kafka
confluent-kafka==2.6.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # also used as FastAPI test client transport

# Dev utilities
python-dotenv==1.0.1

